backend refactors (mostly regarding views)
